+++ src/ReinforcementLearningCore/src/policies/agent/multi_agent.jl
@@ -1,4 +1,4 @@
-export MultiAgentPolicy, MultiAgentHook, Player
+export MultiAgentPolicy, MultiAgentHook, Player, PlayerNamedTuple
 
 using Random # for RandomPolicy
 
@@ -9,6 +9,23 @@ import Base.push!
 struct Player <: AbstractPlayer
     name::Symbol
 end
+
+"""
+    PlayerNamedTuple
+
+A NamedTuple that maps players to their respective values.
+"""
+struct PlayerNamedTuple{N,T}
+    data::NamedTuple{N,T}
+
+    function PlayerNamedTuple(data::Pair...)
+        nt = NamedTuple(first(item).name => last(item) for item in data)
+        new{typeof(nt).parameters...}(nt)
+    end
+end
+
+Base.getindex(nt::PlayerNamedTuple, player::Player) = nt.data[player.name]
+Base.keys(nt::PlayerNamedTuple) = Player.(keys(nt.data))
 
 """
     MultiAgentPolicy(agents::NT) where {NT<: NamedTuple}
