+++ src/ReinforcementLearningCore/src/utils/utils.jl
@@ -2,4 +2,3 @@ include("basic.jl")
 include("stack_frames.jl")
 include("distributions.jl")
 include("networks.jl")
-include("player_named_tuple.jl")
