+++ src/ReinforcementLearningCore/test/policies/multi_agent.jl
@@ -3,6 +3,15 @@ using ReinforcementLearningTrajectories
 using ReinforcementLearningBase
 using DomainSets
 
+
+@testset "Basic PlayerNamedTuple tests" begin
+    nt = PlayerNamedTuple(Player(Symbol(1)) => 3, Player(Symbol(2)) => 4)
+    @test nt.data == (; Symbol(1) => 3, Symbol(2) => 4)
+    @test typeof(nt).parameters == typeof(nt.data).parameters
+    @test nt[Player(Symbol(1))] == 3
+end
+
+
 @testset "MultiAgentPolicy" begin
     trajectory_1 = Trajectory(
         CircularArraySARTSTraces(; capacity = 1),
