+++ src/ReinforcementLearningCore/test/utils/utils.jl
@@ -2,4 +2,3 @@ include("base.jl")
 include("stack_frames.jl")
 include("networks.jl")
 include("distributions.jl")
-include("player_named_tuple.jl")
